+++ tests/OrderTest.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Nevadskiy\Position\Tests\Unit;
+namespace Nevadskiy\Position\Tests;
 
 use Mockery;
 use Nevadskiy\Position\Tests\Support\Factories\CategoryFactory;
 use Nevadskiy\Position\Tests\Support\Models\Category;
-use Nevadskiy\Position\Tests\TestCase;
 
 class OrderTest extends TestCase
 {
@@ -14,11 +13,21 @@ class OrderTest extends TestCase
      */
     public function it_can_order_models_by_position(): void
     {
-        $category2 = CategoryFactory::new()->position(2)->create();
-        $category0 = CategoryFactory::new()->position(0)->create();
-        $category1 = CategoryFactory::new()->position(1)->create();
+        $category2 = CategoryFactory::new()
+            ->position(2)
+            ->create();
 
-        $categories = Category::query()->orderByPosition()->get();
+        $category0 = CategoryFactory::new()
+            ->position(0)
+            ->create();
+
+        $category1 = CategoryFactory::new()
+            ->position(1)
+            ->create();
+
+        $categories = Category::query()
+            ->orderByPosition()
+            ->get();
 
         static::assertTrue($categories[0]->is($category0));
         static::assertTrue($categories[1]->is($category1));
@@ -30,11 +39,21 @@ class OrderTest extends TestCase
      */
     public function it_can_order_models_by_inverse_position(): void
     {
-        $category2 = CategoryFactory::new()->position(2)->create();
-        $category0 = CategoryFactory::new()->position(0)->create();
-        $category1 = CategoryFactory::new()->position(1)->create();
+        $category2 = CategoryFactory::new()
+            ->position(2)
+            ->create();
 
-        $categories = Category::query()->orderByInversePosition()->get();
+        $category0 = CategoryFactory::new()
+            ->position(0)
+            ->create();
+
+        $category1 = CategoryFactory::new()
+            ->position(1)
+            ->create();
+
+        $categories = Category::query()
+            ->orderByInversePosition()
+            ->get();
 
         static::assertTrue($categories[0]->is($category2));
         static::assertTrue($categories[1]->is($category1));
@@ -46,14 +65,24 @@ class OrderTest extends TestCase
      */
     public function it_can_be_ordered_by_position_by_default(): void
     {
-        $category2 = CategoryFactory::new()->position(2)->create();
-        $category0 = CategoryFactory::new()->position(0)->create();
-        $category1 = CategoryFactory::new()->position(1)->create();
+        $category2 = CategoryFactory::new()
+            ->position(2)
+            ->create();
+
+        $category0 = CategoryFactory::new()
+            ->position(0)
+            ->create();
+
+        $category1 = CategoryFactory::new()
+            ->position(1)
+            ->create();
 
         $fakeCategory = Mockery::mock(new Category());
         $fakeCategory->shouldReceive('alwaysOrderByPosition')->andReturnTrue();
 
-        $categories = Category::query()->setModel($fakeCategory)->get();
+        $categories = Category::query()
+            ->setModel($fakeCategory)
+            ->get();
 
         static::assertTrue($categories[0]->is($category0));
         static::assertTrue($categories[1]->is($category1));
@@ -65,9 +94,17 @@ class OrderTest extends TestCase
      */
     public function it_is_not_ordered_by_position_by_default(): void
     {
-        $category2 = CategoryFactory::new()->position(2)->create();
-        $category0 = CategoryFactory::new()->position(0)->create();
-        $category1 = CategoryFactory::new()->position(1)->create();
+        $category2 = CategoryFactory::new()
+            ->position(2)
+            ->create();
+
+        $category0 = CategoryFactory::new()
+            ->position(0)
+            ->create();
+
+        $category1 = CategoryFactory::new()
+            ->position(1)
+            ->create();
 
         $categories = Category::query()->get();