+++ tests/ScopePositionQueryTest.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Nevadskiy\Position\Tests\Unit;
+namespace Nevadskiy\Position\Tests;
 
 use Nevadskiy\Position\Tests\Support\Factories\BookFactory;
 use Nevadskiy\Position\Tests\Support\Factories\CategoryFactory;
-use Nevadskiy\Position\Tests\TestCase;
 
 class ScopePositionQueryTest extends TestCase
 {